xoa tep

+++ controllers/admin.js
@@ -4,6 +4,8 @@ const { validationResult } = require('express-validator/check');
 
 const Product = require('../models/product');
 
+const fileHepler=require('../util/fileHepler')
+
 exports.getAddProduct = (req, res, next) => {
   res.render('admin/edit-product', {
     pageTitle: 'Add Product',
@@ -156,6 +158,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if(image){
+        fileHepler.deleteFile(product.imageUrl)
         product.imageUrl = image.path;
       }
       return product.save().then(result => {
@@ -191,14 +194,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
-    .catch(err => {
-      const error = new Error(err);
-      error.httpStatusCode = 500;
-      return next(error);
-    });
+  Product.findById(prodId)
+    .then(prod=>{
+      if(!prod){
+        return next(new Error('Product not found'))
+      }
+      fileHepler.deleteFile(prod.imageUrl)
+      return Product.deleteOne({ _id: prodId, userId: req.user._id })
+      })
+    .catch(err=>next(err))
+  ;
 };
